requirements.txt
streamlit
openai
python-dotenv
streamlit-mic-recorder